maintenance update
why:
  some handy features were intended to support the unit test from
  the clique/clique_test.go source (the other one is from
  clique/snapshot_test.go.)
  as this test cannot realistically be implemented without the full
  api (includes mining support), it is left as that

+++ tests/test_clique/voter_samples.nim
@@ -29,7 +29,7 @@ type
   TestSpecs* = object   ## Define the various voting scenarios to test
     id*: int                  ## Test id
     info*: string             ## Test description
-    epoch*:   uint64          ## Number of blocks in an epoch (unset = 30000)
+    epoch*: int               ## Number of blocks in an epoch (unset = 30000)
     signers*: seq[string]     ## Initial list of authorized signers in the
                               ## genesis
     votes*: seq[TesterVote]   ## Chain of signed blocks, potentially influencing
